complete order service implementation
at least 90% to work as expected of the original .net application

+++ src/services/basket-service/src/main/java/com/pdt/entities/ShoppingCartItem.java
@@ -1,10 +1,8 @@
 package com.pdt.entities;
 
-import java.math.BigDecimal;
-
 public record ShoppingCartItem(int quantity,
                                String color,
-                               BigDecimal price,
+                               double price,
                                String productId,
                                String productName) {
 }
